requirements.txt
sqlalchemy 
psycopg2-binary
bcrypt
PyJWT
click
fastapi
uvicorn
python-dotenv
python-multipart
requests
rq
alembic
rq-dashboard